Extracted components from the ApiKeys table
